adding authentication feature

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\Api\AgentController;
+use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\CategoryController;
 use App\Http\Controllers\Api\CinemaController;
 use App\Http\Controllers\Api\CityController;
@@ -13,72 +14,68 @@ use App\Http\Controllers\Api\UserTicketController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/user', function (Request $request) {
-    return $request->user();
-})->middleware('auth:sanctum');
-
 
 
 Route::controller(CityController::class)->prefix('cities')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{city}','show');
-    Route::put('/{city}','update');
-    Route::delete('/{city}','destroy');
+    Route::put('/{city}','update')->middleware('auth:sanctum');
+    Route::delete('/{city}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(CinemaController::class)->prefix('cinemas')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{cinema}','show');
-    Route::put('/{cinema}','update');
-    Route::post('/{cinema}/comment','comment');
-    Route::post('/{cinema}/giving_score','givingScore');
-    Route::delete('/{cinema}','destroy');
+    Route::put('/{cinema}','update')->middleware('auth:sanctum');
+    Route::post('/{cinema}/comment','comment')->middleware('auth:sanctum');
+    Route::post('/{cinema}/giving_score','givingScore')->middleware('auth:sanctum');
+    Route::delete('/{cinema}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(PerformanceController::class)->prefix('performances')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{performance}','show');
-    Route::put('/{performance}','update');
-    Route::post('/{performance}/comment','comment');
-    Route::post('/{performance}/giving_score','givingScore');
-    Route::delete('/{performance}','destroy');
+    Route::put('/{performance}','update')->middleware('auth:sanctum');
+    Route::post('/{performance}/comment','comment')->middleware('auth:sanctum');
+    Route::post('/{performance}/giving_score','givingScore')->middleware('auth:sanctum');
+    Route::delete('/{performance}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(DailyScreeningController::class)->prefix('daily_screenings')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{daily_screening}','show');
-    Route::put('/{daily_screening}','update');
-    Route::delete('/{daily_screening}','destroy');
+    Route::put('/{daily_screening}','update')->middleware('auth:sanctum');
+    Route::delete('/{daily_screening}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(AgentController::class)->prefix('agents')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{agent}','show');
-    Route::put('/{agent}','update');
-    Route::delete('/{agent}','destroy');
+    Route::put('/{agent}','update')->middleware('auth:sanctum');
+    Route::delete('/{agent}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(CategoryController::class)->prefix('categories')->group(function () {
     Route::get('/','index');
-    Route::post('/','store');
+    Route::post('/','store')->middleware('auth:sanctum');
     Route::get('/{category}','show');
-    Route::put('/{category}','update');
-    Route::delete('/{category}','destroy');
+    Route::put('/{category}','update')->middleware('auth:sanctum');
+    Route::delete('/{category}','destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(CommentController::class)->prefix('comments')->group(function () {
     Route::get('/','index');
     Route::get('/{comment}','show');
-    Route::post('/{comment}/giving_score','givingScore');
-    Route::delete('/{comment}','destroy');
+    Route::post('/{comment}/giving_score','givingScore')->middleware('auth:sanctum');
+    Route::delete('/{comment}','destroy')->middleware('auth:sanctum');
 });
 
-Route::controller(UserController::class)->prefix('users')->group(function () {
+Route::controller(UserController::class)->middleware('auth:sanctum')->prefix('users')->group(function () {
     Route::get('/','index');
     Route::get('/{user}','show');
     Route::put('/{user}','update');
@@ -86,16 +83,22 @@ Route::controller(UserController::class)->prefix('users')->group(function () {
 });
 
 Route::controller(ScoreController::class)->prefix('scores')->group(function () {
-    Route::get('/','index');
+    Route::get('/','index')->middleware('auth:sanctum');
     Route::get('/{score}','show');
-    Route::delete('/{score}','destroy');
+    Route::delete('/{score}','destroy')->middleware('auth:sanctum');
 });
 
-Route::controller(UserTicketController::class)->prefix('my_tickets')->group(function () {
+Route::controller(UserTicketController::class)->middleware('auth:sanctum')->prefix('my_tickets')->group(function () {
     Route::get('/','index');
     Route::post('/','store');
     Route::get('/{userTicket}','show');
     Route::put('/{userTicket}','update');
     Route::put('/{userTicket}/paid','ticketPaid');
     Route::delete('/{userTicket}','destroy');
+});
+
+Route::controller(AuthController::class)->prefix('auth')->group(function () {
+    Route::post('/login','login');
+    Route::post('/register','register');
+    route::delete('/logout','logout')->middleware('auth:sanctum');
 });
